deepcode_lab/papers/3/generate_code/requirements.txt
torch>=1.7
torchvision>=0.8
tqdm>=4.41
numpy>=1.20
scipy>=1.5
matplotlib>=3.3
pandas>=1.2
pillow>=8.0
pyyaml>=5.3